Cleans up SORT and UNPIVOT factories
